pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slower"
version = "0.1.0"
description = "GPS-based speed limiter for 2006 BMW 325xi (E90 N52) via K+DCAN cable"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Philip Fox Jones", email = "philip@foxpress.design"},
]
dependencies = [
    "pyserial>=3.5",
    "flask>=3.0",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
slower = "slower.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
